+++ docs/CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 <!-- The line "## <square-bracket>Unreleased</square-bracket>" is replaced by the ci/release.py script with the new release version and release date. -->
 
-## [3.18.1-rc.1] - [2021-02-07]
+## [Unreleased]
 
-### Fixed
+### Added
 
-- Prevent traceback when a timeout collecting measurement data occurs. Fixes [#1898](https://github.com/ICTU/quality-time/issues/1898).
+- Add a 'merge requests' metric and add GitLab as source. Closes [#1644](https://github.com/ICTU/quality-time/issues/1644).
 
 ## [3.18.0] - [2021-02-03]